requirements.txt
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26
scipy>=1.11
statsmodels>=0.14
arch>=6.3
cvxpy>=1.4
scikit-learn>=1.4
matplotlib>=3.8
seaborn>=0.13
discord.py>=2.3
pydantic>=2.6
python-dotenv>=1.0
